Rewrite Flow getting-started tutorial
