+++ documentation/Overview.asciidoc
@@ -23,7 +23,7 @@ This is the documentation for Vaadin Flow. It is divided into the following subs
 
 == Introduction and Getting Started
 ** <<introduction/introduction-overview#,Introduction to Vaadin Flow>>
-** <<introduction/tutorial-get-started#,Getting Started>>
+** <<introduction/tutorial-get-started-first-part#,Getting Started>>
 
 == Router and Navigation
 Router is a core concept in Vaadin Flow, enabling Navigation for your website or web app.
