--- tutorial-getting-started/src/main/java/com/vaadin/flow/demo/helloworld/MainView.java
@@ -1,87 +1,0 @@
-/*
- * Copyright 2000-2018 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.demo.helloworld;
-
-import com.vaadin.flow.component.button.Button;
-import com.vaadin.flow.component.dependency.HtmlImport;
-import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.component.icon.Icon;
-import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
-import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import com.vaadin.flow.component.textfield.TextField;
-import com.vaadin.flow.data.value.ValueChangeMode;
-import com.vaadin.flow.router.Route;
-import com.vaadin.flow.theme.Theme;
-import com.vaadin.flow.theme.lumo.Lumo;
-import com.vaadin.flow.tutorial.annotations.CodeFor;
-
-@CodeFor("introduction/tutorial-get-started.asciidoc")
-/**
- * The main view contains a button and a template element.
- */
-@HtmlImport("styles/shared-styles.html")
-@Route("")
-public class MainView extends VerticalLayout {
-    private CustomerService service = CustomerService.getInstance();
-    private Grid<Customer> grid = new Grid<>();
-    private TextField filterText = new TextField();
-    private CustomerForm form = new CustomerForm(this);
-
-    public MainView() {
-        filterText.setPlaceholder("Filter by name...");
-        filterText.setValueChangeMode(ValueChangeMode.EAGER);
-        filterText.addValueChangeListener(e -> updateList());
-
-        Button clearFilterTextBtn = new Button(
-                new Icon(VaadinIcon.CLOSE_CIRCLE));
-        clearFilterTextBtn.addClickListener(e -> filterText.clear());
-
-        HorizontalLayout filtering = new HorizontalLayout(filterText,
-                clearFilterTextBtn);
-
-        Button addCustomerBtn = new Button("Add new customer");
-        addCustomerBtn.addClickListener(e -> {
-            grid.asSingleSelect().clear();
-            form.setCustomer(new Customer());
-        });
-
-        HorizontalLayout toolbar = new HorizontalLayout(filtering,
-                addCustomerBtn);
-
-        grid.setSizeFull();
-
-        grid.addColumn(Customer::getFirstName).setHeader("First name");
-        grid.addColumn(Customer::getLastName).setHeader("Last name");
-        grid.addColumn(Customer::getStatus).setHeader("Status");
-
-        HorizontalLayout main = new HorizontalLayout(grid, form);
-        main.setAlignItems(Alignment.START);
-        main.setSizeFull();
-
-        add(toolbar, main);
-        setHeight("100vh");
-        updateList();
-
-        grid.asSingleSelect().addValueChangeListener(event -> {
-            form.setCustomer(event.getValue());
-        });
-    }
-
-    public void updateList() {
-        grid.setItems(service.findAll(filterText.getValue()));
-    }
-}
