+++ tutorial-getting-started/src/main/java/com/vaadin/starter/skeleton/CustomerStatus.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.demo.helloworld;
+package com.vaadin.starter.skeleton;
 
 import com.vaadin.flow.tutorial.annotations.Helper;
