+++ tutorial-getting-started/src/main/java/com/vaadin/starter/skeleton/ExampleTemplate.java
@@ -13,14 +13,14 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.demo.helloworld;
+package com.vaadin.starter.skeleton;
 
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.HtmlImport;
 import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
-import com.vaadin.flow.demo.helloworld.ExampleTemplate.ExampleModel;
 import com.vaadin.flow.templatemodel.TemplateModel;
 import com.vaadin.flow.tutorial.annotations.CodeFor;
+import com.vaadin.starter.skeleton.ExampleTemplate.ExampleModel;
 
 @CodeFor("introduction/tutorial-get-started.asciidoc")
 /**
